Moving chat classes to 'gui/chat' folder
