+++ src/Common/gui/chat/ChatMessagePanel.cpp
