+++ src/Common/gui/chat/ChatMessagePanel.h
