+++ src/Common/gui/chat/ChatMessagePanel.ui
