+++ src/Common/gui/chat/ChatPanel.cpp
