+++ src/Common/gui/chat/ChatPanel.h
