+++ src/Common/gui/chat/ChatPanel.ui
